+++ packages/contracts/contracts/interfaces/IBlackScholes.sol
@@ -2,35 +2,35 @@
 pragma solidity ^0.8.9;
 
 interface IBlackScholes {
-  struct PricesDeltaStdVega {
-    uint callPrice;
-    uint putPrice;
-    int callDelta;
-    int putDelta;
-    uint stdVega;
-  }
+    struct PricesDeltaStdVega {
+        uint256 callPrice;
+        uint256 putPrice;
+        int256 callDelta;
+        int256 putDelta;
+        uint256 stdVega;
+    }
 
-  function abs(int x) external pure returns (uint);
+    function abs(int256 x) external pure returns (uint256);
 
-  function exp(uint x) external pure returns (uint);
+    function exp(uint256 x) external pure returns (uint256);
 
-  function exp(int x) external pure returns (uint);
+    function exp(int256 x) external pure returns (uint256);
 
-  function sqrt(uint x) external pure returns (uint y);
+    function sqrt(uint256 x) external pure returns (uint256 y);
 
-  function optionPrices(
-    uint timeToExpirySec,
-    uint volatilityDecimal,
-    uint spotDecimal,
-    uint strikeDecimal,
-    int rateDecimal
-  ) external pure returns (uint call, uint put);
+    function optionPrices(
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
+        int256 rateDecimal
+    ) external pure returns (uint256 call, uint256 put);
 
-  function pricesDeltaStdVega(
-    uint timeToExpirySec,
-    uint volatilityDecimal,
-    uint spotDecimal,
-    uint strikeDecimal,
-    int rateDecimal
-  ) external pure returns (PricesDeltaStdVega memory);
+    function pricesDeltaStdVega(
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
+        int256 rateDecimal
+    ) external pure returns (PricesDeltaStdVega memory);
 }
