+++ packages/contracts/contracts/interfaces/IHackMoneyStrategy.sol
@@ -2,16 +2,16 @@
 pragma solidity ^0.8.9;
 
 interface IHackMoneyStrategy {
-    function setBoard(uint boardId) external;
+    function setBoard(uint256 boardId) external;
 
-    function doTrade(uint size)
+    function doTrade(uint256 size)
         external
         returns (
-            uint positionId1,
-            uint positionId2,
-            uint premium,
-            uint collateralAdded,
-            uint premiumExchangeValue
+            uint256 positionId1,
+            uint256 positionId2,
+            uint256 premium,
+            uint256 collateralAdded,
+            uint256 premiumExchangeValue
         );
 
     function returnFundsAndClearStrikes() external;
