+++ packages/contracts/contracts/interfaces/ISynthetix.sol
@@ -2,16 +2,16 @@
 pragma solidity ^0.8.9;
 
 interface ISynthetix {
-  function exchange(
-    bytes32 sourceCurrencyKey,
-    uint sourceAmount,
-    bytes32 destinationCurrencyKey
-  ) external returns (uint amountReceived);
+    function exchange(
+        bytes32 sourceCurrencyKey,
+        uint256 sourceAmount,
+        bytes32 destinationCurrencyKey
+    ) external returns (uint256 amountReceived);
 
-  function exchangeOnBehalf(
-    address exchangeForAddress,
-    bytes32 sourceCurrencyKey,
-    uint sourceAmount,
-    bytes32 destinationCurrencyKey
-  ) external returns (uint amountReceived);
+    function exchangeOnBehalf(
+        address exchangeForAddress,
+        bytes32 sourceCurrencyKey,
+        uint256 sourceAmount,
+        bytes32 destinationCurrencyKey
+    ) external returns (uint256 amountReceived);
 }
